* Updating to use `uint64_t` instead of `size_t` for indices. * Faster random access.   * Iterate once for shallow access (if desired index is in first pass), also measures size.   * From desired index and size, extrapolate iteration value.   * Jump forward to the correct iteration value and iterate through childen (second pass). * Adding `<< "index: " << index`  output for `ASSERT_EQ` in `src/stim/util_top/reference_sample_tree.test.cc` for easier test debugging.

+++ src/stim/util_top/reference_sample_tree.h
@@ -25,7 +25,7 @@ struct ReferenceSampleTree {
     /// Checks if two trees are not exactly the same, including structure (not just uncompressed contents).
     bool operator!=(const ReferenceSampleTree &other) const;
     /// Returns the bit value for a given absolute index.
-    bool operator[](size_t index) const;
+    bool operator[](uint64_t index) const;
     /// Returns a simple description of the tree's structure, like "5*('101'+6*('11'))".
     std::string str() const;
 
@@ -48,7 +48,7 @@ struct ReferenceSampleTree {
     /// Helper method for `simplified`.
     void flatten_and_simplify_into(std::vector<ReferenceSampleTree> &out) const;
     /// Helper method for `operator[]`.
-    bool try_get_bit_value(size_t desired_absolute_index, size_t & current_absolute_index, bool & bit_value) const;
+    bool try_get_bit_value(uint64_t desired_absolute_index, uint64_t & current_absolute_index, bool & bit_value) const;
 };
 std::ostream &operator<<(std::ostream &out, const ReferenceSampleTree &v);
